+++ prov/slurm/helper_test.go
@@ -15,6 +15,7 @@
 package slurm
 
 import (
+	"context"
 	"crypto/rand"
 	"crypto/rsa"
 	"crypto/x509"
@@ -351,10 +352,61 @@ func TestGetJobInfoWithInvalidJob(t *testing.T) {
 	t.Parallel()
 	s := &sshutil.MockSSHClient{
 		MockRunCommand: func(cmd string) (string, error) {
-			return "slurm_load_jobs error: Invalid job id specified", errors.New("")
+			if strings.HasPrefix(cmd, "scontrol") {
+				return "slurm_load_jobs error: Invalid job id specified", errors.New("")
+			}
+			return "", nil
 		},
 	}
-	info, err := getJobInfo(s, "1234")
+	info, err := getJobInfo(context.Background(), s, "d1", "1234")
+	require.Nil(t, info, "info should be nil")
+
+	require.Equal(t, true, isNoJobFoundError(err), "expected no job found error")
+}
+
+func TestGetJobInfoWithEmptyResponseButAccounting(t *testing.T) {
+	t.Parallel()
+	s := &sshutil.MockSSHClient{
+		MockRunCommand: func(cmd string) (string, error) {
+			if strings.HasPrefix(cmd, "scontrol") {
+				return "", nil
+			}
+			return "COMPLETED", nil
+		},
+	}
+	info, err := getJobInfo(context.Background(), s, "d1", "1234")
+	require.NotNil(t, info, "info should not be nil")
+	require.Nil(t, err)
+	require.Equal(t, map[string]string{"JobState": "COMPLETED"}, info, "unexpected job info")
+}
+
+func TestGetJobInfoWithEmptyResponseAndAccountingError(t *testing.T) {
+	t.Parallel()
+	s := &sshutil.MockSSHClient{
+		MockRunCommand: func(cmd string) (string, error) {
+			if strings.HasPrefix(cmd, "scontrol") {
+				return "", nil
+			}
+			return "", errors.New("some error")
+		},
+	}
+	info, err := getJobInfo(context.Background(), s, "d1", "1234")
+	require.Nil(t, info, "info should be nil")
+
+	require.NotNil(t, err, "expecting an error")
+}
+
+func TestGetJobInfoWithEmptyResponseAndAccountingNotConfigured(t *testing.T) {
+	t.Parallel()
+	s := &sshutil.MockSSHClient{
+		MockRunCommand: func(cmd string) (string, error) {
+			if strings.HasPrefix(cmd, "scontrol") {
+				return "", nil
+			}
+			return errMsgAccountingDisabled, errors.New("1")
+		},
+	}
+	info, err := getJobInfo(context.Background(), s, "d1", "1234")
 	require.Nil(t, info, "info should be nil")
 
 	require.Equal(t, true, isNoJobFoundError(err), "expected no job found error")
@@ -374,7 +426,7 @@ func TestGetJobInfo(t *testing.T) {
 	require.Nil(t, err, "unexpected error while opening test file")
 	expected, err := parseJobInfo(data)
 	require.Nil(t, err, "Unexpected error parsing job info")
-	info, err := getJobInfo(s, "1234")
+	info, err := getJobInfo(context.Background(), s, "d1", "1234")
 	require.Nil(t, err, "Unexpected error retrieving job info")
 	require.NotNil(t, info, "info should not be nil")
 
@@ -388,7 +440,7 @@ func TestGetJobInfoWithEmptyResponse(t *testing.T) {
 			return "", nil
 		},
 	}
-	info, err := getJobInfo(s, "1234")
+	info, err := getJobInfo(context.Background(), s, "d1", "1234")
 	require.Nil(t, info, "info should be nil")
 
 	require.Equal(t, true, isNoJobFoundError(err), "expected no job found error")
@@ -401,7 +453,7 @@ func TestGetJobInfoWithError(t *testing.T) {
 			return "oups, it's bad", errors.New("this is an error !")
 		},
 	}
-	info, err := getJobInfo(s, "1234")
+	info, err := getJobInfo(context.Background(), s, "d1", "1234")
 	require.Nil(t, info, "info should be nil")
 
 	require.Equal(t, "oups, it's bad: this is an error !", err.Error(), "expected error")